requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
torch==1.13.0
torchvision==0.14.0
pillow==9.3.0
